Add openocd bridge plus server functionality to tb
Set ENABLE_OPENOCD=1 to enable the correct path in the tb. When starting the
simulator link librbs.so from remote_bitbang/. Afterwards running the simulator
while print out a port. Set the environment variable JTAG_VPI_PORT to it so that
riscv-openocd, when you execute it, knows where to connect to.

This is wip there still seems to be a few bugs lurking around, especially we
don't trigger fetch enable yet. This needs more work in the openocd script.

rtl/tb/remote_bitbang/*:
Produces dynamic library that can be linked in into the simulator. It receives
jtag commands via openocd's bitbang protocol and calls the appropriate jtag DPI
functions that in turn stimulate the pulpissimo soc.

rtl/tb/pulpissimo:
This is the openocd script that properly handles the two jtag taps of
pulpissimo (dmi of the debug unit and custom). The timeouts are set so high
because the simulator is usually very slow.