Fix testbench due to starting off with fetch_enable=1
On the very first instructions the core goes into the bootrom and gets stuck
into an infinite loop/wfi. We then transition out of that by transfering control
to the debugger (halt_harts) and making it go into an infinite loop in
BEGIN_L2_INSTR (by writing dpc and doing halt/resume).