Move testing sequence into jtag_pkg
We move the whole debug module testing code into jtag_pkg to reduce the
clutter in the main testbench file. There is a new parameter
ENABLE_DM_TESTS which can be set to one (default, temporary) to enable
the debug module tests, otherwise we directly start loading the binary
to L2 and execute.