Fix bus access over pulp tap, add option to use it tesbench
Accomodate for the bypass register bit of the debug module tap when
in the read* and write* tasks.

Add option USE_PULP_BUS_ACCESS to prefer the pulp tap for loading the
binary instead of the system bus access of the debug module. This mode
is currently much faster at loading the executable. Enabled by default.