Restructure different tb modes and readd support for flash boot
Flash booting doesn't need and want any configuration so we bypass all
the jtag stuff. What we still need to do is setup system bus access via
the debug module to poll the "end of program" register in 32'h1A1040A0.
For that we moved some parts (set_sbreadonaddr and init_dmi_access) a
bit closer to this final polling loop.

Furthermore we restructured the statements separating the different boot
modes which was a bit convoluted beforehand. This doesn't change
anything.