sim: Add workaround for incisive
We have a two modules with the same name (axi_slice_dc_master_wrap and
axi_slice_dc_slave_wrap) in the axi ip and in pulp_soc/components. The
one is components is outdated and must be removed soon. For now we work
around this issue by removing in manually from the generated file list.
This only affects incisive.